refactor: move navbar js to file & add scrolling
