+++ resources/views/navbar.blade.php
@@ -10,40 +10,7 @@
     ][$breakpoint ?? 'md'];
 @endphp
 
-<div x-data="{
-        open: false,
-        openDropdown: null,
-        selectedChild: null,
-        scrollProgress: 0,
-        nav: null,
-        init() {
-            const { nav } = this.$refs
-            this.nav = nav;
-            window.onscroll = this.onScroll.bind(this);
-            this.scrollProgress = this.getScrollProgress();
-            this.updateShadow(this.scrollProgress);
-        },
-        onScroll() {
-            const progress = this.getScrollProgress()
-            if (progress !== this.scrollProgress) {
-                this.scrollProgress = progress;
-                this.updateShadow(progress);
-            }
-        },
-        getScrollProgress() {
-            const navbarHeight = 82;
-            return Math.min(1, document.documentElement.scrollTop / navbarHeight);
-        },
-        updateShadow(progress) {
-            const maxTransparency = 0.22;
-            const shadowTransparency = Math.round(maxTransparency * progress * 100) / 100;
-            const borderTransparency =  Math.round((1 - progress) * 100) / 100;
-            this.nav.style.boxShadow = `0px 2px 10px 0px rgba(192, 200, 207, ${shadowTransparency})`;
-            this.nav.style.borderColor = `rgba(219, 222, 229, ${borderTransparency})`;
-        }
-    }"
-    x-init="init"
->
+<div x-data="Navbar.dropdown()" x-init="init">
     <div
         x-show="openDropdown !== null || open"
         class="overflow-y-auto fixed inset-0 z-30 opacity-75 bg-theme-secondary-900 {{ $backdropClass }}"
